tests/submission/Infraware/Forms/contents/Forms/fieldset_name.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>FieldSet_name</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
		<fieldset id="input_field" name="input_field">
 		</fieldset>
	</form>
	<script>

		var field = document.getElementById("input_field");

		if (typeof(field.name) == "string") {
			test(function() {
				assert_equals(field.name, "input_field", "name attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("name attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
